fix: 🐛 address PR comments + import framer motion on client
